chore: refactored and split functionality into packages
